requirements.txt
numpy>=1.21.0
pandas>=1.3.0
yfinance>=0.1.70
python-dotenv>=0.19.0
requests>=2.26.0
ta>=0.7.0  # Technical Analysis library
matplotlib>=3.4.0
scikit-learn>=0.24.0
pytest>=6.2.5 